aeroleads-scraper/requirements.txt
selenium==4.25.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
pandas==2.2.2
python-dotenv==1.0.1
